refactor: create AddForm component
