+++ src/components/add-form/styles.css
@@ -6,24 +6,19 @@
   width: 100%;
 }
 
-.item-form {
-  align-items: center;
-  flex-direction: column;
+.time-frame-label {
+  margin-bottom: 8px;
 }
 
 .time-frame {
   display: flex;
   flex-direction: column;
   font-size: 15px;
-  margin-bottom: 10px;
+  margin-bottom: 60px;
   list-style-type: none;
-  padding: 10px 0;
+  padding: 15px;
 }
 
 .item-name {
-  padding: 10px 0;
-}
-
-.time-frame-question {
   padding: 10px 0;
 }
